backend/requirements.txt
flask==2.2.0
docker==5.0.2
requests==2.26.0
PyYAML==6.0
